feat: v1.5.0 - 接口对齐+线程安全+内部质量全面提升
Phase 1 P0 关键修复:
- BaseSource.getLoginInfo/putLoginInfo 改用 AES/ECB/PKCS5Padding 加密
- Book.group Int→Long 匹配 Legado 位掩码
- StrResponse 补齐 code()/message()/errorBody()/raw() 方法
- SearchBook.type 用 BookType 常量
- BookChapter.start/end 改为 Long? 可空
- RuleDataInterface.putVariable 返回 Boolean，支持 bigVariable

Phase 2 P1 接口对齐+质量:
- getLoginJs/getHeaderMap 恢复 substring+lastIndexOf 解析
- BookSource.getTag 去掉 fallback
- BookSource.equals→equal 避免 shadow
- htmlFormat 用 Jsoup 解析保留 img 标签
- toNumChapter 使用 titleNumPattern
- ConcurrentRateLimiter lastAccessMap 用 compute() 线程安全
- CacheManager memoryCache 改 ConcurrentHashMap
- SharedJsScope WeakReference→强引用 LRU(16)
- 新增 AndroidContext.getAndroidId()
- BookType 新增 text 常量

+++ plugin/src/main/kotlin/io/legado/engine/data/Book.kt
@@ -17,7 +17,7 @@ data class Book(
     override var durChapterIndex: Int = 0,
     var durChapterPos: Int = 0,
     var durChapterPage: Int = 0,
-    var group: Int = 0,
+    var group: Long = 0,
     var tag: String? = null,
     var originName: String = "",
     var origin: String = "",
